Update imageview.py
It can now take an image path or folder path as a command line
parameter, and will open those.
The fallback is still data/images/ as before.

New buttons for navigating the dataset live will decrease load time
significantly, and allows for browsing in both directions as well as
skipping up to 1000 images.