Add evaluation tool for splitting data in good/bad
The tool takes as input either a list of predicted bounding boxes and a
list of ground truths, which it will then calculate recall on, or it can
take as input precalculated recall.

It will then output a good_list.txt and a bad_list.txt containing paths
to images.

good_list.txt contains images which we predicted one or more of the
bounding boxes in. bad_list.txt contains images where we predicted none.
good_list.txt and bad_list.txt can be used as input for imageview.py